Projects from the bachelor study added
Group projects were excluded; points:
IJC - DU1: 11,5/15
IJC - DU2: 14/15
ITY3: 14/15
ITY4: 13,5/15
ITY5: 14/15
INP2: 18/20
ISA: 8/15
all other projects were awarded with full points.

Note: ISA project (DHCP monitoring) worked locally, but in the testing
environment it didn't work for some reason (it worked for the project's
extension, though) - this problem was present in about a half of the
projects with this variant of the project. Take care with this solution
for this variant of the project.